Move billing client wrapper to a separate module
