+++ modules/services/payment/src/main/kotlin/au/com/shiftyjelly/pocketcasts/payment/Logger.kt
@@ -1,0 +1,7 @@
+package au.com.shiftyjelly.pocketcasts.payment
+
+interface Logger {
+    fun info(message: String)
+    fun warning(message: String)
+    fun error(message: String, exception: Throwable)
+}
