+++ modules/services/payment/src/main/kotlin/au/com/shiftyjelly/pocketcasts/payment/billing/BillingResultExtension.kt
@@ -1,0 +1,6 @@
+package au.com.shiftyjelly.pocketcasts.payment.billing
+
+import com.android.billingclient.api.BillingClient
+import com.android.billingclient.api.BillingResult
+
+fun BillingResult.isOk() = responseCode == BillingClient.BillingResponseCode.OK
