+++ settings.gradle.kts
@@ -76,6 +76,7 @@ include(":modules:services:images")
 include(":modules:services:localization")
 include(":modules:services:media-noop")
 include(":modules:services:model")
+include(":modules:services:payment")
 include(":modules:services:preferences")
 include(":modules:services:protobuf")
 include(":modules:services:repositories")
